backend/requirements.txt
flask==3.1.1
flask-socketio==5.5.1
flask-cors==5.0.1
python-socketio[client]==5.12.1
eventlet==0.39.1
supabase==2.13.0
python-dotenv==1.1.0
qiskit==1.4.2
qiskit-aer==0.17.0
